% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plots the accidents on the US map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Numeric or integer value representing the state of the US
as per the FARS data.}

\item{year}{Numeric or integer value representing the year of the accidents.}
}
\value{
Returns a plot of the US containing the accidents based on the state
and year.
}
\description{
The function takes the state number and year as an input, and plots the accidents
on the US map.
}
\details{
The state number must be an integer or numerical and must exist
in the FARS data, otherwise the functions terminates with an error. Also
returns an error if the data file for the year input does not exist. Finally,
returns an error if the state or year does not exist in the data set.

The function requires the use of the dplyr, maps and graphics
libraries.
}
\examples{
\dontrun{fars_map_state(state.num = 1, year = 2013)}

}
