% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read file and create a tibble}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string which contains the name of the csv file
that you want to load and exists in the wd.}
}
\value{
This function returns a table with the first ten rows of the csv file
In the form of a dplyr tibble.
}
\description{
This is a simple function that loads the CSV file that you input, if it exists
in the working directory. If the process is successful a tibble will be outputed
to the console.
}
\details{
This function contains a control fork (if) that allows the handling
of inexistent files or bad spelling of the file's name. In the case that the
file doesn't exist the code will stop and a message will appear indicating
that the process was not successful.

This function uses the readr library in order to read rectangular
data and the dplyr library to create a tibble (data.frame).
}
\examples{
\dontrun{fars_read("accident_2013.csv")}

}
