% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Reads month and year from accident files using dplyr}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A vector or list of years in numeric or integer format.}
}
\value{
Returns a list of tibbles with the same number of rows
as the "accident_<year>.csv.bz2" files and two columns (MONTH and
year).
}
\description{
This function accepts as an input a vector or list of years and returns a list
of data frames with MONTH and year columns based on the data loaded from
"accident_<year>.csv.bz2" files.
}
\details{
The files need to exist in the working directory.
Returns NULL and a warning if the file does not exist.

The function requires the use of the dplyr library.
}
\examples{
\dontrun{fars_read_years(2013:2015)}
\dontrun{fars_read_years(list(2013, 2015))}

}
