% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numerical or integer value indicating the year of the data.}
}
\value{
Returns a character string with the format "accident_<year>.csv.bz2".
}
\description{
This function creates a filename for a .csv.bz2 file based on the year that
you input, yielding a string with the format "accident_<year>.csv.bz2".
}
\details{
It requires a numerical or integer input otherwise ends with an error.
}
\examples{
\dontrun{makefilename(2016)}

}
